W-12276263 - Extracting common GraphParser code (part 1)
